Changed the ionogram resampling algorithm to better correspond to the one used in the reference images. Yet it doesn't give quite the same results, but at least they look similar and it doesn't "create" originally nonexistent lines (like the old did, e.g. it used to double some lines).

+++ AIS-Data-Reader/src/main/java/cz/cuni/mff/peckam/ais/AISResultOverlay.java
@@ -81,7 +81,8 @@ public class AISResultOverlay implements ProductOverlay<Boolean, Float, Float, I
             while (period <= ionogram.getMaxColumnValue()) {
                 if (period >= ionogram.getMinColumnValue()) {
                     for (int i = 0; i < 8; i++) {
-                        values.put(new Tuple<>(ionogram.getMinRowValue() + i * ionogram.getRowHeight(), period),
+                        values.put(new Tuple<>((float) (ionogram.getMinRowValue() + i * ionogram.getRowHeight()),
+                                period),
                                 true);
                     }
                 }
